backup-host/requirements.txt
python-daemon==1.5.5
pyOpenSSL==0.12
